Changed numerical data types to bigger ones in inventory and main spider stats

+++ Assets/Code/MainModule/Inventory.cs
@@ -5,6 +5,6 @@ namespace ThroughAThousandEyes.MainModule
 {
     public class Inventory
     {
-        public int Silk;
+        public long Silk;
     }
 }
